fix: #33 홈 api 컴포넌트로 분리
